refactor: move indexer implementation to internal/
