+++ src/indexer/internal/git_detection.rs
