+++ src/indexer/internal/git_state.rs
