+++ src/indexer/internal/navigation_state.rs
