+++ src/indexer/internal/state_machine.rs
